mancache.rb (feature): cache compilation to speed up builds
recompiling typescript files, when the source files from which they're
compiled haven't changed is a needless waste of resourced. This commit
adds a new cache option that makes Jekyll::Typescript::Manager produce
a digest of the input files for a compilation, and only recompile the
site when the digests have changed (otherwise, resourcing the files
already in the build directory).

I've noticed up to a 6 second speed up on builds that don't involve
javascript resources.